fix(profiling): export lock acquire/release call location and variable name (#9615)
We used to export "<filename>:<line_number>" for the name of a lock.
Whenever lock is initialized, we traverse the stack frame to populate
these. However, we can't get the variable name of the lock as we haven't
yet fully initialized and assigned it to a variable.

Now we traverse the frame whenever `_acquire/_release` is called to get
the caller information. Also, when `_acquire` is called for the first
time, we inspect local and global variables in the caller frame to find
the name of this lock variable. With
`DD_PROFILING_LOCK_NAME_INSPECT_DIR` enabled, the lock profiler also
inspects attributes of local and global variables to find the name of
the lock. This behavior can be turned off with the flag.

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)