Simplify checklists in pull request template
Many PRs end up without completed checklists. This updates the checklist to a single item for the both the author and reviewer.

Even though we lose in granularity in the checklist, I think this is only a theoretical regression as I rarely see partially-completed checklists (and when I do, they rarely come with a comment explaining why which makes it impossible to know if it was an intentional omission).