refactor(debugger): make uploader independent
We refactor the implementation of the debugger in order to make the
logs intake uploader an independent component that can be requested by
various sub-products. This way, the enablement of exception replay no
longer requires starting dynamic instrumentation only to have the
uploader running. Instead, each debugger product can now register with
the logs intake uploader, which will then start and stop as required.