Set cache size to 64kb for module/package/file
Current implementation uses the default size of the `cached` decorator
which is 256. This is often (always?) lower than the number of files of
a Python program loaded in memory creating cache churn. This is
problematic for functions like `filename_to_package` since it performs
an expensive operation by calling `_root_module` that walks the
filesystem.

Simple benchmark show a 5000% improvement on cache acces for
`filename_to_package` function.