fix (redis): define result for try/catch
Prior to this change one of the redis_utils functions attempted to call
a function in a try/catch with `result=result` with result being defined
within the try.

This caused "undefined variable" errors to be raised when the exception
was hit.

After this change, result is initialized as None before the try/catch,
preventing the "undefined variable" errors"