fix (tracing): don't report stopiteration errors
There are several integrations which send StopIteration errors as part
of normal operation (say, when we hit the last element of an interable).
We are currently flagging such spans as errors, when its expected
behavior.

After this change, we still generate the spans, but they are no longer
reported as errors.