chore(internal): remove ddsketch Python dependency (#9851)
This PR replaces the Python ddsketch dependency with the libdatadog
implementation of ddsketch.

The implementation/wrapper is fairly straight forward. The majority of
changes were to tests that were either using internal/private attributes
for assertions, or which were asserting on internal implementation
details of DDSketch which should be covered by the implementation's test
suite.

As well, there was a part of datastreams which was using a DDSketch
which did not need it (the DDSketch was never encoded to protobuf, and
all it was doing was tracking count and sum of value). So a simple
`SumCount` helper class was added for the specific limited behavior
needed by that.

## Checklist
- [ ] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)