fix(langchain): support langchain v0.2 [backport #9495 to 2.9] (#9885)
Backports #9495 to 2.9. This PR is technically a feature but
functionally only a fix (removes errors for the new `langchain>=0.2`
version line), so it should be safe to be backported as a fix, as it is
required for other fix PRs to be backported.

Fixes #9368.

Langchain has released v0.2, which now lists `langchain_community` as an
optional dependency. However in langchain v0.1, `langchain_community`
was a required dependency, and used by our langchain integration to
patch embeddings and vectorstores and calculate token information. This
means that for langchain v2, users that do not have
`langchain_community` installed will run into patch failures/errors as
we try to patch something that does not exist.

This PR fixes the issue by now only patching langchain community if it
is available on the user's environment.

This means, moving forward, if a user does not have
`langchain_community` installed and runs langchain versions >= 1.0, the
langchain integration no longer:
- Calculates and tags cost metrics `langchain.tokens.total_cost` or
dogstatsd metrics for openai-related traces (as it depends on the
`langchain_community.get_openai_token_cost_for_model()` helper)
- Patches embeddings calls (as embeddings are dependent on
`langchain_community` modules) with the exception of OpenAI embeddings
calls (which we patch only if `langchain_openai` is available)
- Patches vectorstores similarity search calls (as vectorstores are
dependent on `langchain_community` modules) with the exception of
pinecone vectorstore similarity search calls (which we patch only if
`langchain_pinecone` is available)

Existing tests dependent on `langchain_community` code have been
modified to be skipped if `langchain_community` does not exist. We've
also added a new riot venv that does not include `langchain_community`
to ensure all other functionality remains the same.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)