fix(asm): one click remote config fix (#9874)
This PR fixes a small problem in the Remote Config layer for
ASM_FEATURES product.
After that PR:

On the worker side:

1. if we receive a key in the rc configuration and the value is non
empty, it means the value was changed
2. if we receive a key in the rc configuration and the value is empty,
it means the value was deleted
3. if we don't receive a key in the rc configuration, it means the value
was unchanged.

Previously, if the value was a list, it was already the case, but if the
value was a dictionary, case 2 and 3 were not correctly separated. Now,
an empty dictionary is sent if the key is deleted (case 2).

Also:
- small refactor
- update unit tests appropriately
- increase coverage on ASM_FEATURES unit tests for one click
- this PR will also be validated by new system tests

- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
(cherry picked from commit ebf3fa99ba5cd58a57fa4c5f1950fb6068754645)