chore(release-script): add ability to promote release candidates (#9977)
This change adds to the release script the ability to promote a release
candidate to a stable minor release. This ability is important to ensure
that the content of a minor release is exactly what underwent testing as
a release candidate.

The mechanism used to accomplish this is making the `BASE` environment
parameter to the script work as either a branch name (current behavior)
or a tag name. A few adjustments to release-note gathering and branch
resolution are made to support this. The way to promote a release
candidate to a minor release is simply to specify `BASE=x.x.0rcx` when
invoking this script.

This branch was used to generate this release
https://github.com/DataDog/dd-trace-py/releases/tag/untagged-9f4f7354e2642b01030e,
which exactly matches 2.10.0rc4 in code content.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)