chore(telemetry): queue app-started on first sent payload (#9767)
## Background

The instrumentation telemetry writer is enabled when `ddtrace` is
imported, this is done so all products can begin queuing telemetry
metrics, logs, and configurations immediately. However the
`TelemetryWriter.app_started` is called when the first trace is sent to
the agent (when the AgentWriter is started), this delay allows us to
capture start up errors and configurations.

## Description

- Ensuring `telemetry_writer.app_started()` is called when the first
CIVisibility, Profiling, and LLMOBS payloads are sent to the agent.
Previously `app-started` (and `app-closing`) events were only sent when
the first trace was submitted. With this change `app-started` and
`app-closed` events are sent when tracing is disabled or products are
run with a datadog agent.

- Renames `TelemetryWriter._app_started_event` to
`TelemetryWriter.app_started`. The `_app_started_event` method is not
internal to the telemetry writer. This name change should hopefully make
this method easier to understand.

## Checklist

- [x] The PR description includes an overview of the change
- [x] The PR description articulates the motivation for the change
- [x] The change includes tests OR the PR description describes a
testing strategy
- [x] The PR description notes risks associated with the change, if any
- [x] Newly-added code is easy to change
- [x] The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- [x] The change includes or references documentation updates if
necessary
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Newly-added code is easy to change
- [x] Release note makes sense to a user of the library
- [x] If necessary, author has acknowledged and discussed the
performance implications of this PR as reported in the benchmarks PR
comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)