refactor: implement product plugin interface
We implement a product plugin interface to:

- decouple core sources from product-specific ones;
- encapsulate (re)starting logic in the core component;

This refactor allows us to write the logic required to ensure a proper
(re)start of all the necessary services under various circumstances (e.g.
fork/uWSGI pre-forking model, ...) in a single place in the core part of
the library, instead of replicating the same logic in each product.