chore(profiling): use lock-free queue to manage samples to avoid frequent alloc/free (#10124)
When profiled with [ddprof](https://github.com/DataDog/ddprof), native
profile shows ~525MiB from Allocated Memory view for dd-trace-py.
![Screenshot 2024-08-08 at 1 15
58 PM](https://github.com/user-attachments/assets/4d79cd30-d6d1-4059-9940-24c222ebc872)

This comes from the fact that `SampleManager` allocates a new `Sample`
whenever it needs and frees right away. We can instead have a pool of
`Sample`s to avoid this pattern. These `Sample`s are used by
Stack/Memory/Lock Collectors. StackCollector MemoryCollector runs on
separate threads, and LockCollector is a simple wrapper around python
standard library. So the access to these `Sample`s need to be thread
safe.

#9617 is a draft PR using simple lock guarded vector of `Sample`s to
solve the problem. However, using locks require us to carefully handle
forking and could lead us to a deadlock when a thread holding the lock.
So this PR attempts to avoid that problem by using a lock free queue
available via libdatadog-ffi.

We set the default capacity for pool to be 4. Even though this is a
small number, as Python runs under GIL, only single thread can be run at
a time and Collectors can create a handful number of `Sample` objects.
Even though this is not the case and we need to allocate and free a lot
of `Sample` objects, this PR is slightly better in a sense that we'll
re-use some of the `Sample`s. I'll defer further investigation of
setting default capacity for later.

With this PR, I only see very minimal allocation happening from the
first profile from ddprof, and the Allocated Memory decreases to 292 MiB
from 525MiB.

![Screenshot 2024-08-08 at 2 05
28 PM](https://github.com/user-attachments/assets/61d8a9c9-13b4-447e-9ffc-6921e550e1fe)
 

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: David Sanchez <838104+sanchda@users.noreply.github.com>