feat(llmobs): truncate LLMObs span events that are >1MB in size (#10148)
Events that are >1MB are not supported by the event platform, and will
be truncated automatically in ways that make them no longer processable
by the LLMObs event processor. This change introduces truncation for
span events that exceed that size limitation.

Truncation is performed by dropping the input and output fields.