chore(di): auto-restart uploader on fork
We make the logs uploader inherit from the new forksafe awakeable
periodic service class. This is not simply a refactor, but a change in
behaviour for the debugger. Because we no longer re-create a new
instance of the debugger on fork, any configuration received by the
parent process is automatically propagated to child processes after
fork. This new behaviour is desirable in case of short-lived worker
processes spawned by the parent via fork. This gives a chance to the
child worker processes to actually see the probe configuration, instead
of having to wait for the new configuration to be delivered to them by
the RC client. In fact, this might never be received if the child
process terminates too early.