chore(ci_visibility): install coverage collector during pytest plugin v2 startup (#10268)
Our coverage collector needs to be installed as early as possible
because it requires instrumenting modules as they are loaded.

The v2 plugin is updated to install the plugin during the
`pytest_load_initial_conftests()` hook. This requires enabling
CIVisibility in order to know whether or not to collect coverage based
on the test service settings API response.

The core event handler returning some settings that are not
session-specific (eg: workspace) are updated to fetch settings from the
`CIVisibility` class, rather than from the session settings (as the
session has yet to be discovered). Session discovery remains in
`pytest_sessionstart()` as there is a possibility that final pytest
configuration resolved in `pytest_configure()` ends up disabling the
plugin.

The pytest per-suite coverage test module (`test_coverage_per_suite.py`)
is also moved from the `ci_visibility` suite to the `pytest` suite where
it belongs.

The v2 plugin representation of coverage is pending a rewrite to using
byte arrays instead of the current segments, and as such, the
duplication of segments (seen in the per-suite tests' expected data) is
left unfixed for now.

No release notes are added since the v2 plugin is still unannounced.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)