chore(di): auto-restart uploader on fork (#10251)
We make the logs uploader inherit from the new forksafe awakeable
periodic service class. This is not simply a refactor, but a change in
behaviour for the debugger. Because we no longer re-create a new
instance of the debugger on fork, any configuration received by the
parent process is automatically propagated to child processes after
fork. This new behaviour is desirable in case of short-lived worker
processes spawned by the parent via fork. This gives a chance to the
child worker processes to actually see the probe configuration, instead
of having to wait for the new configuration to be delivered to them by
the RC client. In fact, this might never be received if the child
process terminates too early.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)